+++ lib/src/editor/block_component/heading_block_component/heading_block_component.dart
@@ -22,17 +22,18 @@ class HeadingBlockKeys {
 
 Node headingNode({
   required int level,
+  String? text,
   Delta? delta,
   String? textDirection,
   Attributes? attributes,
 }) {
   assert(level >= 1 && level <= 6);
-  attributes ??= {'delta': (delta ?? Delta()).toJson()};
   return Node(
     type: HeadingBlockKeys.type,
     attributes: {
-      ...attributes,
-      HeadingBlockKeys.level: level,
+      HeadingBlockKeys.delta: (delta ?? (Delta()..insert(text ?? ''))).toJson(),
+      HeadingBlockKeys.level: level.clamp(1, 6),
+      if (attributes != null) ...attributes,
       if (textDirection != null) HeadingBlockKeys.textDirection: textDirection,
     },
   );
@@ -154,15 +155,16 @@ class _HeadingBlockComponentWidgetState
               placeholderText: placeholderText,
               placeholderTextSpanDecorator: (textSpan) => textSpan
                   .updateTextStyle(
-                    placeholderTextStyle,
-                  )
-                  .updateTextStyle(
                     widget.textStyleBuilder?.call(level) ??
                         defaultTextStyle(level),
+                  )
+                  .updateTextStyle(
+                    placeholderTextStyle,
                   ),
               textDirection: textDirection,
               cursorColor: editorState.editorStyle.cursorColor,
               selectionColor: editorState.editorStyle.selectionColor,
+              cursorWidth: editorState.editorStyle.cursorWidth,
             ),
           ),
         ],
@@ -182,6 +184,7 @@ class _HeadingBlockComponentWidgetState
       node: node,
       delegate: this,
       listenable: editorState.selectionNotifier,
+      remoteSelection: editorState.remoteSelections,
       blockColor: editorState.editorStyle.selectionColor,
       supportTypes: const [
         BlockSelectionType.block,