+++ lib/src/editor/editor_component/service/keyboard_service_widget.dart
@@ -1,4 +1,5 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:appflowy_editor/src/editor/editor_component/service/ime/delta_input_on_floating_cursor_update.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -13,9 +14,11 @@ class KeyboardServiceWidget extends StatefulWidget {
     this.commandShortcutEvents = const [],
     this.characterShortcutEvents = const [],
     this.focusNode,
+    this.contentInsertionConfiguration,
     required this.child,
   });
 
+  final ContentInsertionConfiguration? contentInsertionConfiguration;
   final FocusNode? focusNode;
   final List<CommandShortcutEvent> commandShortcutEvents;
   final List<CharacterShortcutEvent> characterShortcutEvents;
@@ -81,6 +84,11 @@ class KeyboardServiceWidgetState extends State<KeyboardServiceWidget>
         action,
         editorState,
       ),
+      onFloatingCursor: (point) => onFloatingCursorUpdate(
+        point,
+        editorState,
+      ),
+      contentInsertionConfiguration: widget.contentInsertionConfiguration,
     );
 
     focusNode = widget.focusNode ?? FocusNode(debugLabel: 'keyboard service');
@@ -126,9 +134,6 @@ class KeyboardServiceWidgetState extends State<KeyboardServiceWidget>
   }
 
   @override
-  KeyEventResult onKey(RawKeyEvent event) => throw UnimplementedError();
-
-  @override
   Widget build(BuildContext context) {
     Widget child = widget.child;
     // if there is no command shortcut event, we don't need to handle hardware keyboard.
@@ -137,7 +142,7 @@ class KeyboardServiceWidgetState extends State<KeyboardServiceWidget>
       // the Focus widget is used to handle hardware keyboard.
       child = Focus(
         focusNode: focusNode,
-        onKey: _onKey,
+        onKeyEvent: _onKeyEvent,
         child: child,
       );
     }
@@ -157,8 +162,9 @@ class KeyboardServiceWidgetState extends State<KeyboardServiceWidget>
   }
 
   /// handle hardware keyboard
-  KeyEventResult _onKey(FocusNode node, RawKeyEvent event) {
-    if (event is! RawKeyDownEvent || !enableShortcuts) {
+  KeyEventResult _onKeyEvent(FocusNode node, KeyEvent event) {
+    if ((event is! KeyDownEvent && event is! KeyRepeatEvent) ||
+        !enableShortcuts) {
       if (textInputService.composingTextRange != TextRange.empty) {
         return KeyEventResult.skipRemainingHandlers;
       }
@@ -231,6 +237,8 @@ class KeyboardServiceWidgetState extends State<KeyboardServiceWidget>
           textCapitalization: TextCapitalization.sentences,
           inputAction: TextInputAction.newline,
           keyboardAppearance: Theme.of(context).brightness,
+          allowedMimeTypes:
+              widget.contentInsertionConfiguration?.allowedMimeTypes ?? [],
         ),
       );
       // disable shortcuts when the IME active