+++ lib/src/editor/editor_component/service/shortcuts/command/paste_command.dart
@@ -15,6 +15,7 @@ final List<CommandShortcutEvent> pasteCommands = [
 ///
 final CommandShortcutEvent pasteCommand = CommandShortcutEvent(
   key: 'paste the content',
+  getDescription: () => AppFlowyEditorL10n.current.cmdPasteContent,
   command: 'ctrl+v',
   macOSCommand: 'cmd+v',
   handler: _pasteCommandHandler,
@@ -23,6 +24,7 @@ final CommandShortcutEvent pasteCommand = CommandShortcutEvent(
 final CommandShortcutEvent pasteTextWithoutFormattingCommand =
     CommandShortcutEvent(
   key: 'paste the content as plain text',
+  getDescription: () => AppFlowyEditorL10n.current.cmdPasteContentAsPlainText,
   command: 'ctrl+shift+v',
   macOSCommand: 'cmd+shift+v',
   handler: _pasteTextWithoutFormattingCommandHandler,
@@ -110,20 +112,44 @@ extension on EditorState {
     final nodes = plainText
         .split('\n')
         .map(
-          (e) => e
+          (paragraph) => paragraph
             ..replaceAll(r'\r', '')
             ..trimRight(),
         )
-        .map((e) {
-          // parse the url content
-          final Attributes attributes = {};
-          if (_hrefRegex.hasMatch(e)) {
-            attributes[AppFlowyRichTextKeys.href] = e;
+        .map((paragraph) {
+          Delta delta = Delta();
+          if (_hrefRegex.hasMatch(paragraph)) {
+            final firstMatch = _hrefRegex.firstMatch(paragraph);
+            if (firstMatch != null) {
+              int startPos = firstMatch.start;
+              int endPos = firstMatch.end;
+              final String? url = firstMatch.group(0);
+              if (url != null) {
+                /// insert the text before the link
+                if (startPos > 0) {
+                  delta.insert(paragraph.substring(0, startPos));
+                }
+
+                /// insert the link
+                delta.insert(
+                  paragraph.substring(startPos, endPos),
+                  attributes: {AppFlowyRichTextKeys.href: url},
+                );
+
+                /// insert the text after the link
+                if (endPos < paragraph.length) {
+                  delta.insert(paragraph.substring(endPos));
+                }
+              }
+            }
+          } else {
+            delta.insert(paragraph);
           }
-          return Delta()..insert(e, attributes: attributes);
+          return delta;
         })
-        .map((e) => paragraphNode(delta: e))
+        .map((paragraph) => paragraphNode(delta: paragraph))
         .toList();
+
     if (nodes.isEmpty) {
       return;
     }