+++ lib/src/editor/selection_menu/selection_menu_service.dart
@@ -2,7 +2,6 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_editor/src/service/default_text_operations/format_rich_text_style.dart';
 import 'package:flutter/material.dart';
 
-// TODO: this file is too long, need to refactor.
 abstract class SelectionMenuService {
   Offset get offset;
   Alignment get alignment;
@@ -129,7 +128,7 @@ class SelectionMenu extends SelectionMenuService {
       },
     );
 
-    Overlay.of(context).insert(_selectionMenuEntry!);
+    Overlay.of(context, rootOverlay: true).insert(_selectionMenuEntry!);
 
     editorState.service.keyboardService?.disable(showCursor: true);
     editorState.service.scrollService?.disable();
@@ -238,7 +237,7 @@ class SelectionMenu extends SelectionMenuService {
 
 final List<SelectionMenuItem> standardSelectionMenuItems = [
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.text,
+    getName: () => AppFlowyEditorL10n.current.text,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'text',
       isSelected: isSelected,
@@ -250,7 +249,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.heading1,
+    getName: () => AppFlowyEditorL10n.current.heading1,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'h1',
       isSelected: isSelected,
@@ -262,7 +261,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.heading2,
+    getName: () => AppFlowyEditorL10n.current.heading2,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'h2',
       isSelected: isSelected,
@@ -274,7 +273,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.heading3,
+    getName: () => AppFlowyEditorL10n.current.heading3,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'h3',
       isSelected: isSelected,
@@ -298,7 +297,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.numberedList,
+    getName: () => AppFlowyEditorL10n.current.numberedList,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'number',
       isSelected: isSelected,
@@ -310,7 +309,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.checkbox,
+    getName: () => AppFlowyEditorL10n.current.checkbox,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'checkbox',
       isSelected: isSelected,
@@ -322,7 +321,7 @@ final List<SelectionMenuItem> standardSelectionMenuItems = [
     },
   ),
   SelectionMenuItem(
-    name: AppFlowyEditorL10n.current.quote,
+    getName: () => AppFlowyEditorL10n.current.quote,
     icon: (editorState, isSelected, style) => SelectionMenuIconWidget(
       name: 'quote',
       isSelected: isSelected,