+++ lib/src/editor_state.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:collection';
 
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_editor/src/editor/editor_component/service/scroll/auto_scroller.dart';
@@ -104,10 +105,16 @@ class EditorState {
   /// The selection of the editor.
   Selection? get selection => selectionNotifier.value;
 
+  /// Remote selection is the selection from other users.
+  final PropertyValueNotifier<List<RemoteSelection>> remoteSelections =
+      PropertyValueNotifier<List<RemoteSelection>>([]);
+
   /// Sets the selection of the editor.
   set selection(Selection? value) {
     // clear the toggled style when the selection is changed.
-    toggledStyle.clear();
+    if (selectionNotifier.value != value) {
+      _toggledStyle.clear();
+    }
 
     selectionNotifier.value = value;
   }
@@ -150,9 +157,7 @@ class EditorState {
   Stream<(TransactionTime, Transaction)> get transactionStream =>
       _observer.stream;
   final StreamController<(TransactionTime, Transaction)> _observer =
-      StreamController.broadcast(
-    sync: true,
-  );
+      StreamController.broadcast(sync: true);
 
   /// Store the toggled format style, like bold, italic, etc.
   /// All the values must be the key from [AppFlowyRichTextKeys.supportToggled].
@@ -161,13 +166,14 @@ class EditorState {
   ///
   /// NOTES: It only works once;
   ///   after the selection is changed, the toggled style will be cleared.
-  final toggledStyle = <String, bool>{};
-  late final toggledStyleNotifier =
-      ValueNotifier<Map<String, bool>>(toggledStyle);
+  UnmodifiableMapView<String, dynamic> get toggledStyle =>
+      UnmodifiableMapView<String, dynamic>(_toggledStyle);
+  final _toggledStyle = Attributes();
+  late final toggledStyleNotifier = ValueNotifier<Attributes>(toggledStyle);
 
-  void updateToggledStyle(String key, bool value) {
-    toggledStyle[key] = value;
-    toggledStyleNotifier.value = {...toggledStyle};
+  void updateToggledStyle(String key, dynamic value) {
+    _toggledStyle[key] = value;
+    toggledStyleNotifier.value = {..._toggledStyle};
   }
 
   final UndoManager undoManager = UndoManager();
@@ -178,7 +184,10 @@ class EditorState {
     return transaction;
   }
 
-  // TODO: only for testing.
+  bool showHeader = false;
+  bool showFooter = false;
+
+  // only used for testing
   bool disableSealTimer = false;
   bool disableRules = false;
 
@@ -281,25 +290,26 @@ class EditorState {
 
     final completer = Completer<void>();
 
-    // broadcast to other users here, before applying the transaction
-    _observer.add((TransactionTime.before, transaction));
+    if (isRemote) {
+      selection = _applyTransactionFromRemote(transaction);
+    } else {
+      // broadcast to other users here, before applying the transaction
+      _observer.add((TransactionTime.before, transaction));
 
-    for (final operation in transaction.operations) {
-      Log.editor.debug('apply op: ${operation.toJson()}');
-      _applyOperation(operation);
-    }
+      _applyTransactionInLocal(transaction);
 
-    // broadcast to other users here, after applying the transaction
-    _observer.add((TransactionTime.after, transaction));
+      // broadcast to other users here, after applying the transaction
+      _observer.add((TransactionTime.after, transaction));
 
-    _recordRedoOrUndo(options, transaction);
+      _recordRedoOrUndo(options, transaction);
 
-    if (withUpdateSelection) {
-      _selectionUpdateReason = SelectionUpdateReason.transaction;
-      if (transaction.selectionExtraInfo != null) {
-        selectionExtraInfo = transaction.selectionExtraInfo;
+      if (withUpdateSelection) {
+        _selectionUpdateReason = SelectionUpdateReason.transaction;
+        if (transaction.selectionExtraInfo != null) {
+          selectionExtraInfo = transaction.selectionExtraInfo;
+        }
+        selection = transaction.afterSelection;
       }
-      selection = transaction.afterSelection;
     }
 
     completer.complete();
@@ -525,18 +535,66 @@ class EditorState {
     });
   }
 
-  void _applyOperation(Operation op) {
-    if (op is InsertOperation) {
-      document.insert(op.path, op.nodes);
-    } else if (op is UpdateOperation) {
-      // ignore the update operation if the attributes are the same.
-      if (!mapEquals(op.attributes, op.oldAttributes)) {
-        document.update(op.path, op.attributes);
+  void _applyTransactionInLocal(Transaction transaction) {
+    for (final op in transaction.operations) {
+      Log.editor.debug('apply op (local): ${op.toJson()}');
+
+      if (op is InsertOperation) {
+        document.insert(op.path, op.nodes);
+      } else if (op is UpdateOperation) {
+        // ignore the update operation if the attributes are the same.
+        if (!mapEquals(op.attributes, op.oldAttributes)) {
+          document.update(op.path, op.attributes);
+        }
+      } else if (op is DeleteOperation) {
+        document.delete(op.path, op.nodes.length);
+      } else if (op is UpdateTextOperation) {
+        document.updateText(op.path, op.delta);
       }
-    } else if (op is DeleteOperation) {
-      document.delete(op.path, op.nodes.length);
-    } else if (op is UpdateTextOperation) {
-      document.updateText(op.path, op.delta);
     }
+  }
+
+  Selection? _applyTransactionFromRemote(Transaction transaction) {
+    var selection = this.selection;
+
+    for (final op in transaction.operations) {
+      Log.editor.debug('apply op (remote): ${op.toJson()}');
+
+      if (op is InsertOperation) {
+        document.insert(op.path, op.nodes);
+        if (selection != null) {
+          if (op.path <= selection.start.path) {
+            selection = Selection(
+              start: selection.start.copyWith(
+                path: selection.start.path.nextNPath(op.nodes.length),
+              ),
+              end: selection.end.copyWith(
+                path: selection.end.path.nextNPath(op.nodes.length),
+              ),
+            );
+          }
+        }
+      } else if (op is UpdateOperation) {
+        document.update(op.path, op.attributes);
+      } else if (op is DeleteOperation) {
+        document.delete(op.path, op.nodes.length);
+        if (selection != null) {
+          if (op.path <= selection.start.path) {
+            selection = Selection(
+              start: selection.start.copyWith(
+                path: selection.start.path.previous,
+              ),
+              end: selection.end.copyWith(
+                path: selection.end.path.previous,
+              ),
+            );
+          }
+        }
+      } else if (op is UpdateTextOperation) {
+        document.updateText(op.path, op.delta);
+      }
+    }
+
+    return selection;
   }
 }