+++ lib/src/plugins/html/html_document_decoder.dart
@@ -421,7 +421,7 @@ class DocumentHTMLDecoder extends Converter<String, Document> {
     if (backgroundColor != null) {
       final highlightColor = backgroundColor.tryToColor()?.toHex();
       if (highlightColor != null) {
-        attributes[AppFlowyRichTextKeys.highlightColor] = highlightColor;
+        attributes[AppFlowyRichTextKeys.backgroundColor] = highlightColor;
       }
     }
 
@@ -430,7 +430,7 @@ class DocumentHTMLDecoder extends Converter<String, Document> {
     if (background != null) {
       final highlightColor = background.tryToColor()?.toHex();
       if (highlightColor != null) {
-        attributes[AppFlowyRichTextKeys.highlightColor] = highlightColor;
+        attributes[AppFlowyRichTextKeys.backgroundColor] = highlightColor;
       }
     }
 
@@ -473,6 +473,7 @@ class HTMLTags {
   static const h1 = 'h1';
   static const h2 = 'h2';
   static const h3 = 'h3';
+  static const br = 'br';
   static const orderedList = 'ol';
   static const unorderedList = 'ul';
   static const list = 'li';